# Route pedestrians over highway=platform