--- a/profiles/foot.lua
+++ b/profiles/foot.lua
@@ -90,6 +90,7 @@ function setup()
         path            = walking_speed,
         steps           = walking_speed,
         pedestrian      = walking_speed,
+        platform        = walking_speed,
         footway         = walking_speed,
         pier            = walking_speed,
       },