Fixes #5410



<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1151.97<br/>plain u32: 1152.58<br/>aliased double: 1159.62<br/>plain double: 1173.39 | aliased u32: 1179.43<br/>plain u32: 1156.27<br/>aliased double: 1196.15<br/>plain double: 1207.88 |
| e2e_match_ch | Ops: 43.48 ± 0.06 ops/s. Best: 43.54 ops/s<br/>Total: 3013.14ms ± 3.97ms. Best: 3008.39ms<br/>Min time: 2.01ms ± 0.03ms<br/>Mean time: 23.00ms ± 0.03ms<br/>Median time: 16.91ms ± 0.08ms<br/>95th percentile: 76.74ms ± 0.35ms<br/>99th percentile: 93.79ms ± 0.68ms<br/>Max time: 106.47ms ± 0.37ms | Ops: 43.06 ± 0.06 ops/s. Best: 43.16 ops/s<br/>Total: 3042.16ms ± 4.02ms. Best: 3034.90ms<br/>Min time: 2.07ms ± 0.03ms<br/>Mean time: 23.22ms ± 0.03ms<br/>Median time: 17.09ms ± 0.16ms<br/>95th percentile: 76.84ms ± 0.35ms<br/>99th percentile: 96.06ms ± 0.40ms<br/>Max time: 109.57ms ± 0.43ms |
| e2e_match_mld | Ops: 64.39 ± 0.10 ops/s. Best: 64.56 ops/s<br/>Total: 2034.42ms ± 3.43ms. Best: 2028.97ms<br/>Min time: 1.70ms ± 0.01ms<br/>Mean time: 15.53ms ± 0.03ms<br/>Median time: 8.01ms ± 0.09ms<br/>95th percentile: 52.29ms ± 0.11ms<br/>99th percentile: 60.67ms ± 0.17ms<br/>Max time: 69.38ms ± 0.37ms | Ops: 63.76 ± 0.41 ops/s. Best: 64.16 ops/s<br/>Total: 2054.92ms ± 13.37ms. Best: 2041.92ms<br/>Min time: 1.74ms ± 0.03ms<br/>Mean time: 15.69ms ± 0.10ms<br/>Median time: 8.28ms ± 0.28ms<br/>95th percentile: 52.18ms ± 0.22ms<br/>99th percentile: 60.92ms ± 0.51ms<br/>Max time: 70.27ms ± 1.14ms |
| e2e_nearest_ch | Ops: 874.39 ± 2.31 ops/s. Best: 879.20 ops/s<br/>Total: 1143.72ms ± 3.04ms. Best: 1137.40ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.14ms ± 0.00ms<br/>Median time: 1.06ms ± 0.00ms<br/>95th percentile: 1.59ms ± 0.00ms<br/>99th percentile: 1.64ms ± 0.01ms<br/>Max time: 5.45ms ± 2.48ms | Ops: 816.36 ± 17.59 ops/s. Best: 834.55 ops/s<br/>Total: 1226.48ms ± 28.85ms. Best: 1198.26ms<br/>Min time: 0.99ms ± 0.00ms<br/>Mean time: 1.23ms ± 0.03ms<br/>Median time: 1.08ms ± 0.01ms<br/>95th percentile: 1.68ms ± 0.06ms<br/>99th percentile: 3.06ms ± 0.63ms<br/>Max time: 8.05ms ± 1.58ms |
| e2e_nearest_mld | Ops: 872.55 ± 2.86 ops/s. Best: 877.51 ops/s<br/>Total: 1146.12ms ± 3.99ms. Best: 1139.59ms<br/>Min time: 0.98ms ± 0.00ms<br/>Mean time: 1.15ms ± 0.00ms<br/>Median time: 1.06ms ± 0.00ms<br/>95th percentile: 1.59ms ± 0.00ms<br/>99th percentile: 1.64ms ± 0.01ms<br/>Max time: 5.75ms ± 2.71ms | Ops: 835.16 ± 20.97 ops/s. Best: 853.73 ops/s<br/>Total: 1197.09ms ± 30.88ms. Best: 1171.33ms<br/>Min time: 0.98ms ± 0.00ms<br/>Mean time: 1.20ms ± 0.03ms<br/>Median time: 1.08ms ± 0.01ms<br/>95th percentile: 1.63ms ± 0.04ms<br/>99th percentile: 2.39ms ± 0.66ms<br/>Max time: 7.16ms ± 1.70ms |
| e2e_route_ch | Ops: 379.11 ± 0.95 ops/s. Best: 380.65 ops/s<br/>Total: 2638.00ms ± 6.70ms. Best: 2627.10ms<br/>Min time: 1.19ms ± 0.01ms<br/>Mean time: 2.64ms ± 0.01ms<br/>Median time: 2.66ms ± 0.00ms<br/>95th percentile: 3.46ms ± 0.02ms<br/>99th percentile: 3.85ms ± 0.03ms<br/>Max time: 6.35ms ± 2.14ms | Ops: 347.18 ± 9.65 ops/s. Best: 358.84 ops/s<br/>Total: 2883.89ms ± 84.31ms. Best: 2786.79ms<br/>Min time: 1.22ms ± 0.02ms<br/>Mean time: 2.88ms ± 0.09ms<br/>Median time: 2.81ms ± 0.05ms<br/>95th percentile: 4.12ms ± 0.32ms<br/>99th percentile: 5.93ms ± 0.80ms<br/>Max time: 8.80ms ± 1.25ms |
| e2e_route_mld | Ops: 317.94 ± 0.84 ops/s. Best: 318.81 ops/s<br/>Total: 3145.39ms ± 8.58ms. Best: 3136.64ms<br/>Min time: 1.18ms ± 0.02ms<br/>Mean time: 3.15ms ± 0.01ms<br/>Median time: 3.18ms ± 0.01ms<br/>95th percentile: 4.27ms ± 0.02ms<br/>99th percentile: 4.76ms ± 0.09ms<br/>Max time: 7.36ms ± 2.21ms | Ops: 297.80 ± 4.48 ops/s. Best: 303.88 ops/s<br/>Total: 3356.43ms ± 50.24ms. Best: 3290.82ms<br/>Min time: 1.19ms ± 0.00ms<br/>Mean time: 3.36ms ± 0.05ms<br/>Median time: 3.35ms ± 0.03ms<br/>95th percentile: 4.69ms ± 0.14ms<br/>99th percentile: 6.05ms ± 0.40ms<br/>Max time: 8.90ms ± 1.05ms |
| e2e_table_ch | Ops: 317.89 ± 0.50 ops/s. Best: 318.81 ops/s<br/>Total: 3145.77ms ± 5.21ms. Best: 3136.62ms<br/>Min time: 1.61ms ± 0.02ms<br/>Mean time: 3.15ms ± 0.01ms<br/>Median time: 3.14ms ± 0.02ms<br/>95th percentile: 4.38ms ± 0.02ms<br/>99th percentile: 4.69ms ± 0.03ms<br/>Max time: 7.95ms ± 2.19ms | Ops: 314.43 ± 4.85 ops/s. Best: 318.78 ops/s<br/>Total: 3183.03ms ± 53.31ms. Best: 3136.96ms<br/>Min time: 1.59ms ± 0.03ms<br/>Mean time: 3.18ms ± 0.05ms<br/>Median time: 3.14ms ± 0.05ms<br/>95th percentile: 4.47ms ± 0.12ms<br/>99th percentile: 6.04ms ± 0.80ms<br/>Max time: 10.20ms ± 1.68ms |
| e2e_table_mld | Ops: 111.81 ± 0.12 ops/s. Best: 112.01 ops/s<br/>Total: 8943.79ms ± 10.13ms. Best: 8927.84ms<br/>Min time: 3.57ms ± 0.05ms<br/>Mean time: 8.94ms ± 0.01ms<br/>Median time: 8.92ms ± 0.05ms<br/>95th percentile: 13.73ms ± 0.04ms<br/>99th percentile: 14.36ms ± 0.05ms<br/>Max time: 16.81ms ± 2.19ms | Ops: 107.51 ± 0.97 ops/s. Best: 109.07 ops/s<br/>Total: 9303.75ms ± 92.57ms. Best: 9168.25ms<br/>Min time: 3.70ms ± 0.10ms<br/>Mean time: 9.30ms ± 0.09ms<br/>Median time: 9.28ms ± 0.10ms<br/>95th percentile: 14.20ms ± 0.16ms<br/>99th percentile: 15.36ms ± 0.56ms<br/>Max time: 19.19ms ± 2.97ms |
| e2e_trip_ch | Ops: 100.77 ± 1.03 ops/s. Best: 102.28 ops/s<br/>Total: 9920.63ms ± 96.31ms. Best: 9777.51ms<br/>Min time: 1.43ms ± 0.08ms<br/>Mean time: 9.92ms ± 0.10ms<br/>Median time: 9.36ms ± 0.12ms<br/>95th percentile: 17.98ms ± 0.16ms<br/>99th percentile: 19.90ms ± 0.28ms<br/>Max time: 25.95ms ± 6.21ms | Ops: 96.22 ± 0.73 ops/s. Best: 97.16 ops/s<br/>Total: 10396.07ms ± 79.59ms. Best: 10291.89ms<br/>Min time: 1.53ms ± 0.10ms<br/>Mean time: 10.39ms ± 0.08ms<br/>Median time: 9.81ms ± 0.06ms<br/>95th percentile: 18.53ms ± 0.12ms<br/>99th percentile: 20.83ms ± 0.26ms<br/>Max time: 25.83ms ± 1.55ms |
| e2e_trip_mld | Ops: 59.48 ± 0.53 ops/s. Best: 60.46 ops/s<br/>Total: 16812.92ms ± 150.63ms. Best: 16541.05ms<br/>Min time: 1.58ms ± 0.24ms<br/>Mean time: 16.81ms ± 0.16ms<br/>Median time: 16.32ms ± 0.16ms<br/>95th percentile: 27.67ms ± 0.21ms<br/>99th percentile: 29.97ms ± 0.24ms<br/>Max time: 32.31ms ± 0.64ms | Ops: 57.12 ± 0.62 ops/s. Best: 58.00 ops/s<br/>Total: 17518.22ms ± 189.37ms. Best: 17242.49ms<br/>Min time: 1.67ms ± 0.24ms<br/>Mean time: 17.51ms ± 0.18ms<br/>Median time: 17.03ms ± 0.23ms<br/>95th percentile: 28.60ms ± 0.29ms<br/>99th percentile: 31.22ms ± 0.42ms<br/>Max time: 36.35ms ± 3.24ms |
| json-render | String: 5.65948ms<br/>Stringstream: 9.07366ms<br/>Vector: 6.64167ms | String: 5.88992ms<br/>Stringstream: 9.48467ms<br/>Vector: 7.0287ms |
| match_ch | Default radius: <br/>4.60445ms/req at 82 coordinate<br/>0.0561518ms/coordinate<br/>Radius 10m: <br/>16.1309ms/req at 82 coordinate<br/>0.196718ms/coordinate | Default radius: <br/>4.64504ms/req at 82 coordinate<br/>0.0566468ms/coordinate<br/>Radius 10m: <br/>16.3167ms/req at 82 coordinate<br/>0.198984ms/coordinate |
| match_mld | Default radius: <br/>2.97131ms/req at 82 coordinate<br/>0.0362354ms/coordinate<br/>Radius 10m: <br/>11.0419ms/req at 82 coordinate<br/>0.134658ms/coordinate | Default radius: <br/>3.18335ms/req at 82 coordinate<br/>0.0388214ms/coordinate<br/>Radius 10m: <br/>11.1431ms/req at 82 coordinate<br/>0.135892ms/coordinate |
| osrm_contract | Time: 98.07s Peak RAM: 202.08MB | Time: 97.84s Peak RAM: 200.95MB |
| osrm_customize | Time: 1.30s Peak RAM: 117.69MB | Time: 1.32s Peak RAM: 117.80MB |
| osrm_extract | Time: 11.81s Peak RAM: 433.24MB | Time: 11.76s Peak RAM: 430.28MB |
| osrm_partition | Time: 2.10s Peak RAM: 143.43MB | Time: 2.11s Peak RAM: 142.93MB |
| packedvector | random write:<br/>std::vector 9968.55 ms<br/>util::packed_vector 74583.9 ms<br/>slowdown: 7.48192<br/>random read:<br/>std::vector 8635.45 ms<br/>util::packed_vector 31130.2 ms<br/>slowdown: 3.60494 | random write:<br/>std::vector 11276.2 ms<br/>util::packed_vector 73443.2 ms<br/>slowdown: 6.51314<br/>random read:<br/>std::vector 12142.1 ms<br/>util::packed_vector 30380.3 ms<br/>slowdown: 2.50205 |
| random_match_ch | 500 matches, default radius<br/>ops: 199.64 ± 1.02 ops/s. best: 200.73ops/s.<br/>total: 285.52 ± 1.48ms. best: 283.97ms.<br/>avg: 5.01 ± 0.03ms<br/>min: 0.13 ± 0.01ms<br/>max: 25.94 ± 0.10ms<br/>p99: 25.94 ± 0.10ms<br/><br/>500 matches, radius=10<br/>ops: 58.82 ± 0.09 ops/s. best: 59.01ops/s.<br/>total: 1088.15 ± 1.74ms. best: 1084.57ms.<br/>avg: 17.00 ± 0.03ms<br/>min: 0.15 ± 0.00ms<br/>max: 251.18 ± 0.70ms<br/>p99: 251.18 ± 0.70ms<br/><br/>500 matches, radius=20<br/>ops: 14.00 ± 0.02 ops/s. best: 14.03ops/s.<br/>total: 4644.12 ± 7.28ms. best: 4633.15ms.<br/>avg: 71.45 ± 0.11ms<br/>min: 0.30 ± 0.01ms<br/>max: 1272.56 ± 3.84ms<br/>p99: 1272.56 ± 3.84ms | 500 matches, default radius<br/>ops: 202.27 ± 0.96 ops/s. best: 203.45ops/s.<br/>total: 281.81 ± 1.41ms. best: 280.17ms.<br/>avg: 4.94 ± 0.02ms<br/>min: 0.14 ± 0.01ms<br/>max: 26.43 ± 0.20ms<br/>p99: 26.43 ± 0.20ms<br/><br/>500 matches, radius=10<br/>ops: 59.89 ± 0.12 ops/s. best: 60.14ops/s.<br/>total: 1068.71 ± 2.23ms. best: 1064.15ms.<br/>avg: 16.70 ± 0.03ms<br/>min: 0.15 ± 0.01ms<br/>max: 244.64 ± 0.62ms<br/>p99: 244.64 ± 0.62ms<br/><br/>500 matches, radius=20<br/>ops: 14.38 ± 0.03 ops/s. best: 14.43ops/s.<br/>total: 4520.31 ± 10.86ms. best: 4503.14ms.<br/>avg: 69.54 ± 0.17ms<br/>min: 0.31 ± 0.00ms<br/>max: 1227.23 ± 4.42ms<br/>p99: 1227.23 ± 4.42ms |
| random_match_mld | 500 matches, default radius<br/>ops: 303.93 ± 1.34 ops/s. best: 304.99ops/s.<br/>total: 187.55 ± 0.83ms. best: 186.89ms.<br/>avg: 3.29 ± 0.01ms<br/>min: 0.12 ± 0.01ms<br/>max: 18.91 ± 0.04ms<br/>p99: 18.91 ± 0.04ms<br/><br/>500 matches, radius=10<br/>ops: 107.38 ± 0.31 ops/s. best: 107.84ops/s.<br/>total: 596.03 ± 1.73ms. best: 593.45ms.<br/>avg: 9.31 ± 0.03ms<br/>min: 0.14 ± 0.00ms<br/>max: 110.74 ± 0.62ms<br/>p99: 110.74 ± 0.62ms<br/><br/>500 matches, radius=20<br/>ops: 21.76 ± 0.03 ops/s. best: 21.80ops/s.<br/>total: 2987.46 ± 4.65ms. best: 2982.29ms.<br/>avg: 45.96 ± 0.07ms<br/>min: 0.19 ± 0.00ms<br/>max: 581.44 ± 1.64ms<br/>p99: 581.44 ± 1.64ms | 500 matches, default radius<br/>ops: 302.98 ± 1.63 ops/s. best: 304.49ops/s.<br/>total: 188.14 ± 1.02ms. best: 187.20ms.<br/>avg: 3.30 ± 0.02ms<br/>min: 0.12 ± 0.01ms<br/>max: 18.90 ± 0.02ms<br/>p99: 18.90 ± 0.02ms<br/><br/>500 matches, radius=10<br/>ops: 107.10 ± 0.30 ops/s. best: 107.66ops/s.<br/>total: 597.57 ± 1.62ms. best: 594.48ms.<br/>avg: 9.34 ± 0.03ms<br/>min: 0.14 ± 0.00ms<br/>max: 110.83 ± 0.41ms<br/>p99: 110.83 ± 0.41ms<br/><br/>500 matches, radius=20<br/>ops: 21.74 ± 0.03 ops/s. best: 21.79ops/s.<br/>total: 2990.53 ± 4.36ms. best: 2983.49ms.<br/>avg: 46.01 ± 0.07ms<br/>min: 0.20 ± 0.00ms<br/>max: 583.48 ± 1.04ms<br/>p99: 583.48 ± 1.04ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 23174.76 ± 54.14 ops/s. best: 23220.17ops/s.<br/>total: 431.51 ± 1.01ms. best: 430.66ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.05ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18015.09 ± 10.55 ops/s. best: 18028.21ops/s.<br/>total: 555.09 ± 0.32ms. best: 554.69ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14489.97 ± 18.22 ops/s. best: 14518.44ops/s.<br/>total: 690.13 ± 0.87ms. best: 688.78ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.20 ± 0.03ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 23032.64 ± 196.74 ops/s. best: 23205.96ops/s.<br/>total: 434.22 ± 3.82ms. best: 430.92ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.18 ± 0.06ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17754.07 ± 87.67 ops/s. best: 17867.40ops/s.<br/>total: 563.27 ± 2.78ms. best: 559.68ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14376.32 ± 41.53 ops/s. best: 14415.58ops/s.<br/>total: 695.60 ± 2.02ms. best: 693.69ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 23140.65 ± 126.90 ops/s. best: 23303.96ops/s.<br/>total: 432.16 ± 2.38ms. best: 429.11ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17887.25 ± 15.94 ops/s. best: 17910.39ops/s.<br/>total: 559.06 ± 0.50ms. best: 558.34ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14426.75 ± 11.80 ops/s. best: 14440.10ops/s.<br/>total: 693.16 ± 0.57ms. best: 692.52ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 22778.12 ± 60.57 ops/s. best: 22875.62ops/s.<br/>total: 439.02 ± 1.23ms. best: 437.15ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.05ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17859.75 ± 20.56 ops/s. best: 17897.25ops/s.<br/>total: 559.92 ± 0.65ms. best: 558.75ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14351.63 ± 24.69 ops/s. best: 14386.13ops/s.<br/>total: 696.79 ± 1.20ms. best: 695.11ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 529.71 ± 3.12 ops/s. best: 532.94ops/s.<br/>total: 1857.72 ± 10.98ms. best: 1846.36ms.<br/>avg: 1.89 ± 0.01ms<br/>min: 0.30 ± 0.00ms<br/>max: 3.87 ± 1.17ms<br/>p99: 2.80 ± 0.24ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 584.31 ± 8.57 ops/s. best: 591.20ops/s.<br/>total: 1712.03 ± 25.76ms. best: 1691.46ms.<br/>avg: 1.71 ± 0.03ms<br/>min: 0.05 ± 0.00ms<br/>max: 5.30 ± 0.99ms<br/>p99: 3.86 ± 0.20ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 987.00 ± 10.99 ops/s. best: 1010.43ops/s.<br/>total: 997.13 ± 10.99ms. best: 973.85ms.<br/>avg: 1.01 ± 0.01ms<br/>min: 0.26 ± 0.00ms<br/>max: 1.69 ± 0.05ms<br/>p99: 1.45 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 966.57 ± 32.78 ops/s. best: 1017.59ops/s.<br/>total: 1035.98 ± 32.91ms. best: 982.72ms.<br/>avg: 1.04 ± 0.03ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.32 ± 0.48ms<br/>p99: 2.58 ± 0.10ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 517.26 ± 10.46 ops/s. best: 527.30ops/s.<br/>total: 1903.49 ± 39.59ms. best: 1866.10ms.<br/>avg: 1.93 ± 0.04ms<br/>min: 0.30 ± 0.01ms<br/>max: 4.89 ± 1.56ms<br/>p99: 2.84 ± 0.09ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 601.00 ± 5.43 ops/s. best: 607.39ops/s.<br/>total: 1664.06 ± 15.11ms. best: 1646.40ms.<br/>avg: 1.66 ± 0.02ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.64 ± 0.06ms<br/>p99: 3.55 ± 0.05ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1009.54 ± 21.54 ops/s. best: 1029.92ops/s.<br/>total: 975.28 ± 21.20ms. best: 955.41ms.<br/>avg: 0.99 ± 0.02ms<br/>min: 0.26 ± 0.00ms<br/>max: 1.59 ± 0.03ms<br/>p99: 1.44 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1103.47 ± 3.41 ops/s. best: 1108.41ops/s.<br/>total: 906.24 ± 2.80ms. best: 902.19ms.<br/>avg: 0.91 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 3.65 ± 0.02ms<br/>p99: 2.26 ± 0.02ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 260.74 ± 0.74 ops/s. best: 261.34ops/s.<br/>total: 3773.88 ± 10.69ms. best: 3765.23ms.<br/>avg: 3.84 ± 0.01ms<br/>min: 0.29 ± 0.01ms<br/>max: 8.46 ± 0.01ms<br/>p99: 6.41 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 252.52 ± 0.32 ops/s. best: 253.02ops/s.<br/>total: 3960.10 ± 4.95ms. best: 3952.18ms.<br/>avg: 3.96 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 9.15 ± 0.24ms<br/>p99: 8.18 ± 0.03ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 338.57 ± 3.04 ops/s. best: 341.71ops/s.<br/>total: 2906.62 ± 26.23ms. best: 2879.65ms.<br/>avg: 2.95 ± 0.03ms<br/>min: 0.28 ± 0.00ms<br/>max: 7.12 ± 0.05ms<br/>p99: 5.06 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 311.63 ± 0.47 ops/s. best: 312.30ops/s.<br/>total: 3208.99 ± 4.83ms. best: 3202.02ms.<br/>avg: 3.21 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 6.83 ± 0.07ms<br/>p99: 6.44 ± 0.04ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 254.14 ± 0.40 ops/s. best: 254.81ops/s.<br/>total: 3871.84 ± 6.03ms. best: 3861.65ms.<br/>avg: 3.93 ± 0.01ms<br/>min: 0.30 ± 0.01ms<br/>max: 8.60 ± 0.05ms<br/>p99: 6.58 ± 0.07ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 245.21 ± 0.35 ops/s. best: 245.76ops/s.<br/>total: 4078.19 ± 5.80ms. best: 4068.94ms.<br/>avg: 4.08 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 10.12 ± 1.04ms<br/>p99: 8.44 ± 0.03ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 336.67 ± 0.50 ops/s. best: 337.21ops/s.<br/>total: 2922.73 ± 4.40ms. best: 2918.03ms.<br/>avg: 2.97 ± 0.00ms<br/>min: 0.27 ± 0.00ms<br/>max: 7.08 ± 0.03ms<br/>p99: 5.07 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 299.81 ± 3.07 ops/s. best: 305.08ops/s.<br/>total: 3335.81 ± 31.56ms. best: 3277.85ms.<br/>avg: 3.34 ± 0.03ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.14 ± 0.12ms<br/>p99: 6.65 ± 0.12ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1376.23 ± 9.62 ops/s. best: 1383.71ops/s.<br/>total: 181.67 ± 1.29ms. best: 180.67ms.<br/>avg: 0.73 ± 0.01ms<br/>min: 0.46 ± 0.01ms<br/>max: 1.09 ± 0.25ms<br/>p99: 0.89 ± 0.03ms<br/><br/>250 tables, 25 coordinates<br/>ops: 163.52 ± 0.30 ops/s. best: 163.80ops/s.<br/>total: 1528.91 ± 2.82ms. best: 1526.21ms.<br/>avg: 6.12 ± 0.01ms<br/>min: 5.25 ± 0.01ms<br/>max: 7.00 ± 0.40ms<br/>p99: 6.85 ± 0.31ms<br/><br/>250 tables, 50 coordinates<br/>ops: 80.43 ± 0.66 ops/s. best: 80.99ops/s.<br/>total: 3108.62 ± 25.79ms. best: 3086.89ms.<br/>avg: 12.43 ± 0.10ms<br/>min: 11.53 ± 0.06ms<br/>max: 17.02 ± 5.04ms<br/>p99: 15.87 ± 3.84ms | 250 tables, 3 coordinates<br/>ops: 1433.50 ± 11.37 ops/s. best: 1444.59ops/s.<br/>total: 174.41 ± 1.40ms. best: 173.06ms.<br/>avg: 0.70 ± 0.01ms<br/>min: 0.46 ± 0.01ms<br/>max: 1.04 ± 0.24ms<br/>p99: 0.88 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 172.16 ± 0.51 ops/s. best: 173.27ops/s.<br/>total: 1452.13 ± 4.30ms. best: 1442.87ms.<br/>avg: 5.81 ± 0.02ms<br/>min: 4.94 ± 0.02ms<br/>max: 6.51 ± 0.04ms<br/>p99: 6.39 ± 0.03ms<br/><br/>250 tables, 50 coordinates<br/>ops: 84.37 ± 0.70 ops/s. best: 85.21ops/s.<br/>total: 2963.56 ± 24.86ms. best: 2933.95ms.<br/>avg: 11.85 ± 0.10ms<br/>min: 10.72 ± 0.05ms<br/>max: 13.28 ± 0.28ms<br/>p99: 12.89 ± 0.19ms |
…